refactor: split server.js to app.js
